refactor(bootstrap): move init-tx builder to hydrozoa.bootstrap
Relocate InitializationTx.Build to hydrozoa.bootstrap.InitializationTxBuilder (producing core's InitializationTx); the running head keeps only InitializationTx.Parse. InitializationTxSeq.Build delegates to it. No sbt submodule or type duplication. (just lint && just fmt applied.) GUM-129 step F (option a).
